// AdguardMini/sciter-ui/modules/common/apis/types/Telemetry.ts
export class TelemetryEvent extends pb_1.Message {
    #one_of_decls: number[][] = [[1, 2]];
    constructor(data?: any[] | ({} & (({
        pageView?: PageView;
        customEvent?: never;
    } | {
        pageView?: never;
        customEvent?: CustomTelemetryEvent;
    })))) {
        super();
        pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
        if (!Array.isArray(data) && typeof data == "object") {
            if ("pageView" in data && data.pageView != undefined) {
                this.pageView = data.pageView;
            }
            if ("customEvent" in data && data.customEvent != undefined) {
                this.customEvent = data.customEvent;
            }
        }
    }
    get pageView() {
        return pb_1.Message.getWrapperField(this, PageView, 1) as PageView;
    }
    set pageView(value: PageView) {
        pb_1.Message.setOneofWrapperField(this, 1, this.#one_of_decls[0], value);
    }
    get has_page_view() {
        return pb_1.Message.getField(this, 1) != null;
    }
    get customEvent() {
        return pb_1.Message.getWrapperField(this, CustomTelemetryEvent, 2) as CustomTelemetryEvent;
    }
    set customEvent(value: CustomTelemetryEvent) {
        pb_1.Message.setOneofWrapperField(this, 2, this.#one_of_decls[0], value);
    }
    get has_custom_event() {
        return pb_1.Message.getField(this, 2) != null;
    }
    get kind() {
        const cases: {
            [index: number]: "none" | "pageView" | "customEvent";
        } = {
            0: "none",
            1: "pageView",
            2: "customEvent"
        };
        return cases[pb_1.Message.computeOneofCase(this, [1, 2])];
    }
    static fromObject(data: {
        pageView?: ReturnType<typeof PageView.prototype.toObject>;
        customEvent?: ReturnType<typeof CustomTelemetryEvent.prototype.toObject>;
    }): TelemetryEvent {
        const message = new TelemetryEvent({});
        if (data.pageView != null) {
            message.pageView = PageView.fromObject(data.pageView);
        }
        if (data.customEvent != null) {
            message.customEvent = CustomTelemetryEvent.fromObject(data.customEvent);
        }
        return message;
    }
    toObject() {
        const data: {
            pageView?: ReturnType<typeof PageView.prototype.toObject>;
            customEvent?: ReturnType<typeof CustomTelemetryEvent.prototype.toObject>;
        } = {};
        if (this.pageView != null) {
            data.pageView = this.pageView.toObject();
        }
        if (this.customEvent != null) {
            data.customEvent = this.customEvent.toObject();
        }
        return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
        const writer = w || new pb_1.BinaryWriter();
        if (this.has_page_view)
            writer.writeMessage(1, this.pageView, () => this.pageView.serialize(writer));
        if (this.has_custom_event)
            writer.writeMessage(2, this.customEvent, () => this.customEvent.serialize(writer));
        if (!w)
            return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): TelemetryEvent {
        const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new TelemetryEvent();
        while (reader.nextField()) {
            if (reader.isEndGroup())
                break;
            switch (reader.getFieldNumber()) {
                case 1:
                    reader.readMessage(message.pageView, () => message.pageView = PageView.deserialize(reader));
                    break;
                case 2:
                    reader.readMessage(message.customEvent, () => message.customEvent = CustomTelemetryEvent.deserialize(reader));
                    break;
                default: reader.skipField();
            }
        }
        return message;
    }
    serializeBinary(): Uint8Array {
        return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): TelemetryEvent {
        return TelemetryEvent.deserialize(bytes);
    }
}
export class PageView extends pb_1.Message {
    #one_of_decls: number[][] = [[2]];
    constructor(data?: any[] | ({
        name?: string;
    } & (({
        refName?: string;
    })))) {
        super();
        pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
        if (!Array.isArray(data) && typeof data == "object") {
            if ("name" in data && data.name != undefined) {
                this.name = data.name;
            }
            if ("refName" in data && data.refName != undefined) {
                this.refName = data.refName;
            }
        }
    }
    get name() {
        return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
    }
    set name(value: string) {
        pb_1.Message.setField(this, 1, value);
    }
    get refName() {
        return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
    }
    set refName(value: string) {
        pb_1.Message.setOneofField(this, 2, this.#one_of_decls[0], value);
    }
    get has_ref_name() {
        return pb_1.Message.getField(this, 2) != null;
    }
    get _refName() {
        const cases: {
            [index: number]: "none" | "refName";
        } = {
            0: "none",
            2: "refName"
        };
        return cases[pb_1.Message.computeOneofCase(this, [2])];
    }
    static fromObject(data: {
        name?: string;
        refName?: string;
    }): PageView {
        const message = new PageView({});
        if (data.name != null) {
            message.name = data.name;
        }
        if (data.refName != null) {
            message.refName = data.refName;
        }
        return message;
    }
    toObject() {
        const data: {
            name?: string;
            refName?: string;
        } = {};
        if (this.name != null) {
            data.name = this.name;
        }
        if (this.refName != null) {
            data.refName = this.refName;
        }
        return data;
    }
    serialize(): Uint8Array;
    serialize(w: pb_1.BinaryWriter): void;
    serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
        const writer = w || new pb_1.BinaryWriter();
        if (this.name.length)
            writer.writeString(1, this.name);
        if (this.has_ref_name)
            writer.writeString(2, this.refName);
        if (!w)
            return writer.getResultBuffer();
    }
    static deserialize(bytes: Uint8Array | pb_1.BinaryReader): PageView {
        const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new PageView();
        while (reader.nextField()) {
            if (reader.isEndGroup())
                break;
            switch (reader.getFieldNumber()) {
                case 1:
                    message.name = reader.readString();
                    break;
                case 2:
                    message.refName = reader.readString();
                    break;
                default: reader.skipField();
            }
        }
        return message;
    }
    serializeBinary(): Uint8Array {
        return this.serialize();
    }
    static deserializeBinary(bytes: Uint8Array): PageView {
        return PageView.deserialize(bytes);
    }
}